poäng och liv

+++ gameobject.h
@@ -86,8 +86,8 @@ public:
     void ship_movement();
     void level_up();
 
-    int life;
-    int point;
+    long life;
+    long point;
     bool lostLife = false;
 };
